requirements_.txt
astra-toolbox==2.3.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
matplotlib==3.9.2
seaborn==0.13.2
pandas
numpy
